refactor(challenge): my-challenge/section 타입버킷 해체 → 목적(feature)별 승격
section/은 "섹션"이라는 타입으로 여러 목적을 묶은 버킷이라 응집도를 떨어뜨림.
목적별 feature 폴더로 승격하고 section/ 제거:
- mission-submit/  mission-guide/  → section 밖으로 (sections+hooks+list-form 통째)
- doc-upload/ (구 my-doc-upload) recommended-program/ 신설
- 평면 섹션들을 소유 feature로: LinkInput/Review/DocumentUpload/PersonalInfoConsent→mission-submit,
  MissionHeader→mission-guide, MissionCalendarSection→mission-calendar,
  MissionMentorComment/OtherMission→mission
- 깨진 경로는 @/ 절대경로로 갱신

이제 my-challenge 직속은 전부 "무엇을 하는가(목적)" 폴더. tsc 0 / lint 통과.
