+++ apps/web/src/app/(user)/challenge/[applicationId]/[programId]/user/info/page.tsx
@@ -10,7 +10,7 @@ import { usePatchUser, useUserQuery } from '@/api/user/user';
 import { AsyncBoundary } from '@/common/boundary/AsyncBoundary';
 import LoadingContainer from '@/common/loading/LoadingContainer';
 import Input from '@/common/input/v2/Input';
-import { DASHBOARD_FIRST_VISIT_GOAL } from '@/domain/challenge/my-challenge/section/mission-submit/MissionSubmitZeroSection';
+import { DASHBOARD_FIRST_VISIT_GOAL } from '@/domain/challenge/my-challenge/mission-submit/MissionSubmitZeroSection';
 import CareerInfoForm, {
   CareerInfoSelections,
   CareerInfoValues,
