+++ apps/web/src/domain/challenge/my-challenge/doc-upload/DocumentFileItem.tsx
