+++ apps/web/src/domain/challenge/my-challenge/mission-guide/MissionGuideBonusSection.tsx
@@ -7,7 +7,7 @@ import { clsx } from 'clsx';
 import { Dayjs } from 'dayjs';
 import { ReactNode } from 'react';
 import MissionGuideSkeleton from './MissionGuideSkeleton';
-import MissionHeaderSection from '../MissionHeaderSection';
+import MissionHeaderSection from '@/domain/challenge/my-challenge/mission-guide/MissionHeaderSection';
 
 const Heading3 = ({
   children,
