+++ apps/web/src/domain/challenge/my-challenge/mission-guide/MissionGuideSkeleton.tsx
