+++ apps/web/src/domain/challenge/my-challenge/mission-guide/MissionHeaderSection.tsx
