+++ apps/web/src/domain/challenge/my-challenge/mission-submit/DocumentUploadSection.tsx
@@ -5,7 +5,7 @@ import { getFileNameFromUrl } from '@/utils/getFileNameFromUrl';
 import { clsx } from 'clsx';
 import { Trash2, Upload } from 'lucide-react';
 import { useRef, type RefObject } from 'react';
-import { UploadedFiles } from './mission-submit/MissionSubmitTalentPoolSection';
+import { UploadedFiles } from '@/domain/challenge/my-challenge/mission-submit/MissionSubmitTalentPoolSection';
 
 const MAX_FILE_SIZE = 50 * 1024 * 1024; // 50MB
