+++ apps/web/src/domain/challenge/my-challenge/mission-submit/MissionSubmitZeroSection.tsx
