+++ apps/web/src/domain/challenge/my-challenge/mission-submit/mission-submit-list-form/components/MissionSubmitGuidance.tsx
