+++ apps/web/src/domain/challenge/my-challenge/mission/OtherMissionSection.tsx
