+++ apps/web/src/domain/challenge/my-challenge/recommended-program/RecommendedProgramSection.tsx
