+++ apps/web/src/domain/challenge/old/OldMyChallengeDashboard.tsx
@@ -1,5 +1,5 @@
 import { useOldCurrentChallenge } from '@/context/OldCurrentChallengeProvider';
-import OtherMissionSection from '@/domain/challenge/my-challenge/section/OtherMissionSection';
+import OtherMissionSection from '@/domain/challenge/my-challenge/mission/OtherMissionSection';
 import OldMissionCalendarSection from '@/domain/challenge/old/OldMissionCalendarSection';
 import OldMyDailyMissionSection from '@/domain/challenge/old/OldMyDailyMissionSection';
 import dayjs from '@/lib/dayjs';
